radeon: Set first parameter of drmOpen() to NULL
Since the introduction of a call to drmCheckModesettingSupported()
in radeon_kernel_mode_enabled(), with abort condition if such call
fails, the first argument to drmOpen() call in radeon_open_drm_master()
has become a no-op red herring.

Such argument (a kernel module name) is supposed to result in load
of specified kernel module. However, this will never happen. The
problem is that if the code containing drmOpen() call is reached, it
means that drmCheckModesettingSupported() check has previously
succeeded, which implies the module is already loaded.

So, drmOpen() will never load a kernel module. But it gives the
impression that it will.

In order to avoid this confusion, this commit replaces it with NULL,
like xf86-video-intel driver does.

Bugzilla: https://bugs.freedesktop.org/show_bug.cgi?id=72849
Signed-off-by: Robert Millan <rmh@freebsd.org>