radeon/exa: Always use a scratch surface for UTS to vram
If we don't, we may hit a buffer that crosses the
visible vram boundary resulting in a sigbus when the
CPU accesses the buffer beyond the PCI aperture.
This will introduce an extra copy in certain cases.

This is based on Michel's patch from bug 44099 updated
for all asic families.

Bug:
https://bugs.freedesktop.org/show_bug.cgi?id=44099

Signed-off-by: Alex Deucher <alexander.deucher@amd.com>
Reviewed-by: Michel Dänzer <michel.daenzer@amd.com>