present: Fall back to modeset for unflip operation
It's not always possible to use the page flip ioctl for this, e.g.
during DPMS off. We were previously just skipping the unflip in that
case, which could result in hangs when setting DPMS off while a
fullscreen Present app is running, e.g. at the GNOME3 lock screen.

Reviewed-by: Alex Deucher <alexander.deucher@amd.com>