Fix inconsistent default eg_tile_split in evergreen_accel.c
eg_tile_split() maps the default: switch-case to tile split
flag 6, like for a tile split size of 4096 Bytes. All other
instances of tile split mappings or reverse mappings in both
the ddx and in Mesa's radeon gallium drivers assign the default:
case to tile split flag 4, consistent with a default of 1024
Bytes.

Make this one outlier consistent with other instances of
mappings in the code to avoid surprises.

Signed-off-by: Mario Kleiner <mario.kleiner.de@gmail.com>