Also handle disabled CRTCs in drmmode_clear_pending_flip
If disabling a CRTC had to be deferred due to a pending flip in
drmmode_crtc_dpms, there may no longer be any outputs associated with
the CRTC when we get here. So we have to check for !crtc->enabled and
call drmmode_crtc_dpms in that case as well.

Fixes: 9090309e057d ("Wait for pending flips to complete before turning
off an output or CRTC")
Reviewed-by: Alex Deucher <alexander.deucher@amd.com>
(Cherry picked from commit 4bd2d01552f18153afa03a8947b22eebf3d67c6b)