Always allow Present page flipping with TearFree
Even if TearFree is active for the the CRTC we're synchronizing to. In
that case, for Present flips synchronized to vertical blank, the other
scanout buffer is immediately synchronized and flipped to during the
target vertical blank period. For Present flips not synchronized to
vertical blank, we simply use the MSC and timestamp values of the last
vertical blank period for timing purposes, and let the normal TearFree
mechanism handle display updates.

v2:
* Move manpage hunk to next change, since TearFree can still prevent
  DRI2 page flipping with this change.

Reviewed-by: Alex Deucher <alexander.deucher@amd.com> # v1